requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-dotenv>=0.19.0
PyYAML>=5.4.1
requests>=2.26.0
pandas>=1.3.0
tqdm>=4.62.0
logging

# RAG, LLM, and Vector DB
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-chroma>=0.0.1
langchain-huggingface>=0.0.1
langchain-unstructured>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
huggingface-hub>=0.16.0
unstructured>=0.10.0
PyPDF2>=3.0.0
docx2txt>=0.8
python-docx>=0.8.11
python-pptx>=0.6.21
python-magic>=0.4.27
python-magic-bin>=0.4.14
qdrant-client>=1.1.0
faiss-cpu>=1.7.4
sentencepiece>=0.1.99
protobuf>=4.24.0

# Utilities
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
tiktoken>=0.4.0
spacy>=3.5.0
nltk>=3.8.0

# Web UI
streamlit>=1.22.0
streamlit-extras>=0.2.0
plotly>=5.13.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
cryptography>=41.0.0
ratelimit>=2.2.1

# Testing
pytest>=7.0.0
httpx>=0.24.0

# Optional: for metrics, monitoring, and cache
redis>=4.5.0
aioredis>=2.0.0
structlog>=23.1.0
python-json-logger>=2.0.0
sentry-sdk>=1.25.0

# Remove all unrelated ML, CV, audio, geospatial, and other unused packages